+++ Sources/CollectionViewModel.swift
@@ -167,3 +167,26 @@ public struct CollectionSectionViewModel: DiffableViewModel {
         self.diffingKey = diffingKey
     }
 }
+
+/// `Collection` support for diffing
+extension CollectionSectionViewModel: Collection {
+    /// :nodoc:
+    public subscript(position: Int) -> CollectionCellViewModel {
+        return self.cellViewModels[position]
+    }
+
+    /// :nodoc:
+    public func index(after i: Int) -> Int {
+        return self.cellViewModels.index(after: i)
+    }
+
+    /// :nodoc:
+    public var startIndex: Int {
+        return self.cellViewModels.startIndex
+    }
+
+    /// :nodoc:
+    public var endIndex: Int {
+        return self.cellViewModels.endIndex
+    }
+}